pyproject.toml
[tool.poetry]
name = "adjust"
version = "0.1.0"
description = ""
authors = ["Jesus Lopez <jesus@jesusla.com>"]
readme = "README.md"

[tool.poetry.scripts]
adjust = "adjust.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "^6.0"
requests = "^2.31.0"
pydantic = "^1.10.8"
gitignore-parser = "^0.1.3"
lxml = "^4.9.2"
beautifulsoup4 = "^4.12.2"
click = "^8.1.7"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
click-help-colors = "^0.9.4"


[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.10"
types-requests = "^2.31.0.0"
types-beautifulsoup4 = "^4.12.0.20240511"
types-tqdm = "^4.66.0.20240417"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
requests-mock = "^1.12.1"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
requests_mock_case_sensitive = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 132

[tool.ruff]
line-length = 132
